Added reverse, made doubly linked list, and personalized clicking to change color
Changed the nodes to include a previous pointer, and implemented a
reverse method which switched the prev and next data members of each
node, reversing the list.  Then I called the reverse method before and
after I drew the rings to the screen, that way the first ring is drawn
last… making it appear as if it were on top.  Then I adde a onThisRing
method which finds the first ring that a mouseClick touches so I can
change just its color.